refactor(analysis): extract standards-text, source-gathering and MCP arg builders
