+++ tools/import_baseline.txt
@@ -1,7 +1,7 @@
 # Grandfathered layer-import violations. Do NOT add entries without
 # justification -- the goal is to burn this list down, not grow it.
 # Regenerate intentionally: python tools/check_imports.py --update-baseline
-src/quodeq/analysis/subprocess.py:283:llm_bridge
+src/quodeq/analysis/_api_source_gathering.py:25:llm_bridge
 src/quodeq/api/routes_project_list.py:227:analysis
 src/quodeq/ci/review.py:134:_cli_evaluation
 src/quodeq/data/fs/repo_clone.py:19:context
