+++ tools/size_baseline.txt
@@ -8,8 +8,6 @@ src/quodeq/_cli_evaluation.py:467:function
 src/quodeq/_cli_evaluation.py:574:function
 src/quodeq/_cli_resolution.py:1:file
 src/quodeq/_cli_resolution.py:253:function
-src/quodeq/analysis/_command.py:1:file
-src/quodeq/analysis/_command.py:73:function
 src/quodeq/analysis/_dim_estimates.py:50:function
 src/quodeq/analysis/_pipeline.py:118:function
 src/quodeq/analysis/api_prompt_assembly.py:76:function
@@ -32,8 +30,6 @@ src/quodeq/analysis/run_lifecycle.py:1:file
 src/quodeq/analysis/subagents/_git_scoring.py:48:function
 src/quodeq/analysis/subagents/_pool_launcher.py:116:function
 src/quodeq/analysis/subagents/_source_files.py:10:function
-src/quodeq/analysis/subprocess.py:1:file
-src/quodeq/analysis/subprocess.py:402:function
 src/quodeq/api/_assistant_helpers.py:190:function
 src/quodeq/api/_evaluation_routes.py:103:function
 src/quodeq/api/_evaluation_routes.py:130:function
